# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Appointment Management System API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/accounts/verify/{token}": {
            "get": {
                "description": "Verifies an account using a token provided in the URL path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Verify Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Verification token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/verify.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/custom_errors.CustomError"
                        }
                    },
                    "422": {
                        "description": "Failed to verify account",
                        "schema": {
                            "$ref": "#/definitions/custom_errors.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/tenants/": {
            "post": {
                "description": "register a new tenant in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "register Tenant",
                "parameters": [
                    {
                        "description": "Tenant registration request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/register.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/register.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/custom_errors.CustomError"
                        }
                    },
                    "422": {
                        "description": "Failed to register tenant",
                        "schema": {
                            "$ref": "#/definitions/custom_errors.CustomError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "custom_errors.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "register.Account": {
            "type": "object",
            "required": [
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "email": {
                    "description": "Admin's email",
                    "type": "string",
                    "example": "admin@example.com"
                },
                "name": {
                    "description": "Admin's name",
                    "type": "string",
                    "example": "John Doe"
                },
                "phone": {
                    "description": "Admin's phone",
                    "type": "string",
                    "example": "+1234567890"
                }
            }
        },
        "register.Request": {
            "description": "Tenant registration request body",
            "type": "object",
            "required": [
                "account",
                "closing_hours",
                "email",
                "name",
                "opening_hours",
                "phone",
                "timezone"
            ],
            "properties": {
                "account": {
                    "description": "Admin user details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/register.Account"
                        }
                    ]
                },
                "address": {
                    "description": "Tenant's address",
                    "type": "string",
                    "maxLength": 255,
                    "example": "123 Main Street"
                },
                "closing_hours": {
                    "description": "Closing hours",
                    "type": "string",
                    "example": "18:00"
                },
                "email": {
                    "description": "Tenant's email",
                    "type": "string",
                    "example": "tenant@example.com"
                },
                "name": {
                    "description": "Tenant's name",
                    "type": "string",
                    "example": "Example Tenant"
                },
                "opening_hours": {
                    "description": "Opening hours",
                    "type": "string",
                    "example": "09:00"
                },
                "phone": {
                    "description": "Tenant's phone",
                    "type": "string",
                    "example": "+1987654321"
                },
                "timezone": {
                    "description": "Tenant's timezone",
                    "type": "string",
                    "example": "America/New_York"
                }
            }
        },
        "register.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "verify.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}